Moved new classes to a separate file
