+++ Source/Utils/Make.package
@@ -4,5 +4,6 @@ CEXE_sources += WarpXUtil.cpp
 CEXE_sources += WarpXAlgorithmSelection.cpp
 CEXE_sources += Average.cpp
 CEXE_sources += Interpolate.cpp
+CEXE_sources += IntervalsParser.cpp
 
 VPATH_LOCATIONS   += $(WARPX_HOME)/Source/Utils
